nvrtc 12 major specifier: convert to int instead of float

+++ build-system/luxmake/wheel.py
@@ -141,7 +141,7 @@ def make_wheel(args):
         # Export METADATA file
         with open(dist_info / "METADATA", "w", encoding="utf-8") as f:
             nvrtc_version = get_dep_version("nvrtc")
-            major = float(nvrtc_version.split('.')[0])
+            major = int(nvrtc_version.split('.')[0])
             requirement = (
                 f"nvidia-cuda-nvrtc-cu{major}=={nvrtc_version}"
                 if major <= 12
